Create maze.java
You need to start from the user given position(input) in the matrix and then check whether the maze has an exit or not. Travelling the maze will be done through the open spaces inside the matrix and if finally led to an open space in the boundary of the matrix.....Then the maze has an "exit" else "no exit there".